rootful: unset XDG_RUNTIME_DIR
Depending how the user logs in to the root account, XDG_RUNTIME_DIR is
set to /run/user/0 or it is unset. For conmon we already set it always
to an empty string. The inconsistency is causing issues for the dnsname
plugin. To fix it unset XDG_RUNTIME_DIR for the podman process.

[NO TESTS NEEDED]

Fixes #10806
Fixes #10745

Signed-off-by: Paul Holzinger <pholzing@redhat.com>