[CI:DOCS] Fix multi-arch image docs
The automation workflow was altered in recent history to build images
daily, even if the podman version didn't change.  This was is necessary
so that any updates/security vulnerabilities in ancillary packages are
incorporated quickly.  However, documentation was never updated to
reflect this change.  This commit puts the two in sync.

Signed-off-by: Chris Evich <cevich@redhat.com>