Use Make GVPROXY_PATH to locate gvproxy
Presented as an alternative to PR #11449

Rather than do backflips in the code to locate `gvproxy`, use a
build-time variable to set the location.  That variable can default to
`/usr/libexec` but other build packages (_e.g._ Homebrew) can set it to
something of their liking.

I'll take no offense if the consensus is that we do not want to pollute
the build, but we should likely add a runtime configuration parameter as an
alternative in that case.

Fixes: #11226

Signed-off-by: Jonathan Springer <jonpspri@gmail.com>